+++ kloch/__init__.py
@@ -1,24 +1,24 @@
 __all__ = [
-    "add_profile_location",
     "config",
     "deepmerge_dicts",
     "EnvironmentProfile",
+    "Environ",
     "filesyntax",
     "get_all_profile_file_paths",
-    "get_profile_locations",
     "get_profile_file_path",
     "get_cli",
     "get_config",
-    "KENV_PROFILE_PATH_ENV_VAR",
     "KlochConfig",
     "launchers",
     "MergeableDict",
     "MergeRule",
     "read_profile_from_file",
+    "read_profile_from_id",
     "refacto_dict",
     "serialize_profile",
     "write_profile_to_file",
 ]
+from .constants import Environ
 from ._dictmerge import MergeableDict
 from ._dictmerge import MergeRule
 from ._dictmerge import refacto_dict
@@ -29,15 +29,13 @@ from . import config
 from . import launchers
 from . import filesyntax
 from .filesyntax import EnvironmentProfile
-from .filesyntax import KENV_PROFILE_PATH_ENV_VAR
 from .filesyntax import serialize_profile
 from .filesyntax import read_profile_from_file
+from .filesyntax import read_profile_from_id
 from .filesyntax import write_profile_to_file
-from .filesyntax import add_profile_location
-from .filesyntax import get_profile_locations
 from .filesyntax import get_profile_file_path
 from .filesyntax import get_all_profile_file_paths
 from .cli import get_cli
 
 # keep in sync with pyproject.toml
-__version__ = "0.7.0"
+__version__ = "0.8.0"