+++ kloch/__main__.py
@@ -1,30 +1,49 @@
 import logging
+import logging.handlers
 import sys
+from typing import List
+from typing import Optional
 
 import kloch
 
 LOGGER = logging.getLogger(__name__)
 
 
-def main():
+def main(argv: Optional[List[str]] = None):
+    """
+    Args:
+        argv: command line arguments. from sys.argv if not provided
+    """
     config = kloch.get_config()
-    cli = kloch.get_cli()
+    cli = kloch.get_cli(argv, config=config)
     log_level = logging.DEBUG if cli.debug else config.cli_logging_default_level
-    logging.basicConfig(
-        level=log_level,
-        format=config.cli_logging_format,
-        style="{",
-        stream=sys.stdout,
-    )
+
+    formatter = logging.Formatter(config.cli_logging_format, style="{")
+
+    logging.root.setLevel(logging.DEBUG)
+
+    handler = logging.StreamHandler(stream=sys.stdout)
+    handler.setLevel(log_level)
+    handler.setFormatter(formatter)
+    logging.root.addHandler(handler)
+
+    for log_path in config.cli_logging_paths:
+        handler = logging.handlers.RotatingFileHandler(
+            log_path,
+            # keep in sync with config option
+            maxBytes=65536,
+            backupCount=1,
+            encoding="utf-8",
+        )
+        handler.setLevel(logging.DEBUG)
+        handler.setFormatter(formatter)
+        logging.root.addHandler(handler)
+
     LOGGER.debug(f"starting {kloch.__name__} v{kloch.__version__}")
     LOGGER.debug(f"retrieved cli with args={cli._args}")
-
-    if cli.profile_paths:
-        LOGGER.debug(f"adding {len(cli.profile_paths)} profile locations")
-        [kloch.add_profile_location(path) for path in cli.profile_paths]
 
     sys.exit(cli.execute())
 
 
 if __name__ == "__main__":
-    main()
+    main()  # pragma: no cover