+++ kloch/filesyntax/__init__.py
@@ -3,11 +3,10 @@ Serialization and unserialization of environment profile to disk.
 """
 
 from ._profile import EnvironmentProfile
-from ._io import KENV_PROFILE_PATH_ENV_VAR
-from ._io import add_profile_location
-from ._io import get_profile_locations
+from ._io import is_file_environment_profile
 from ._io import get_profile_file_path
 from ._io import get_all_profile_file_paths
 from ._io import read_profile_from_file
+from ._io import read_profile_from_id
 from ._io import serialize_profile
 from ._io import write_profile_to_file