+++ tests/test_cli.py
@@ -23,8 +23,17 @@ def test__getCli__subcommands():
     assert isinstance(cli, kloch.cli.ListParser)
 
 
+def test__getCli__list__default(capsys):
+    argv = ["list"]
+    cli = kloch.get_cli(argv=argv)
+    cli.execute()
+
+    captured = capsys.readouterr()
+    assert "Searching 0 locations" in captured.out
+
+
 def test__getCli__list(monkeypatch, data_dir, capsys):
-    monkeypatch.setenv(kloch.KENV_PROFILE_PATH_ENV_VAR, str(data_dir))
+    monkeypatch.setenv(kloch.Environ.CONFIG_PROFILE_PATHS, str(data_dir))
 
     argv = ["list"]
     cli = kloch.get_cli(argv=argv)
@@ -39,7 +48,7 @@ def test__getCli__list(monkeypatch, data_dir, capsys):
 
 
 def test__getCli__list__filter(monkeypatch, data_dir, capsys):
-    monkeypatch.setenv(kloch.KENV_PROFILE_PATH_ENV_VAR, str(data_dir))
+    monkeypatch.setenv(kloch.Environ.CONFIG_PROFILE_PATHS, str(data_dir))
 
     name_filter = ".*es:beta"
     argv = ["list", name_filter]
@@ -50,6 +59,19 @@ def test__getCli__list__filter(monkeypatch, data_dir, capsys):
     assert "Searching 1 locations" in captured.out
     assert f"Filter <{name_filter}> specified" in captured.out
     assert "Found 1 valid profiles" in captured.out
+
+
+def test__getCli__list__profile_paths(data_dir, capsys):
+    argv = ["list", "--profile_paths", str(data_dir)]
+    cli = kloch.get_cli(argv=argv)
+    cli.execute()
+
+    captured = capsys.readouterr()
+    assert "Searching 1 locations" in captured.out
+
+    profile_capture = re.search(r"Found (\d+) valid profiles", captured.out)
+    assert profile_capture
+    assert int(profile_capture.group(1)) >= 1
 
 
 def test__getCli__run(monkeypatch, data_dir):
@@ -64,7 +86,7 @@ def test__getCli__run(monkeypatch, data_dir):
         Results.env = env
         return subprocess.CompletedProcess(command, 0)
 
-    monkeypatch.setenv(kloch.KENV_PROFILE_PATH_ENV_VAR, str(data_dir))
+    monkeypatch.setenv(kloch.Environ.CONFIG_PROFILE_PATHS, str(data_dir))
     monkeypatch.setattr(subprocess, "run", patched_subprocess)
 
     argv = ["run", "lxm"]
@@ -89,7 +111,7 @@ def test__getCli__run__command(monkeypatch, data_dir):
         Results.env = env
         return subprocess.CompletedProcess(command, 0)
 
-    monkeypatch.setenv(kloch.KENV_PROFILE_PATH_ENV_VAR, str(data_dir))
+    monkeypatch.setenv(kloch.Environ.CONFIG_PROFILE_PATHS, str(data_dir))
     monkeypatch.setattr(subprocess, "run", patched_subprocess)
 
     extra_command = ["echo", "a", "bunch", "of", "ÅÍÎÏ˝ÓÔÒÚÆ☃", "--debug"]
@@ -135,7 +157,6 @@ def test__getCli__plugins__undefined(data_dir, capsys):
 
 
 def test__getCli__plugins(monkeypatch, data_dir, capsys):
-
     plugin_path = data_dir / "plugins-behr"
     monkeypatch.syspath_prepend(plugin_path)
     plugin_path = data_dir / "plugins-tyfa"
@@ -156,7 +177,6 @@ def test__getCli__plugins(monkeypatch, data_dir, capsys):
 
 
 def test__getCli__plugins__arg__launcher_plugin(monkeypatch, data_dir, capsys):
-
     plugin_path = data_dir / "plugins-behr"
     monkeypatch.syspath_prepend(plugin_path)
     plugin_path = data_dir / "plugins-tyfa"